Adding project structure
